+++ NEWS.md
@@ -5,7 +5,7 @@
 -   `spsOption` has a new argument `.list`: set more than one options a time by passing a list of options and their values to it.
 -   New function `simpleStack`, a simple stack data structure with methods in R6 class. Methods include `pop`, `push` and more.
 -   New function `historyStack`, a stack data structure to store history steps in R6 class. Similar to browser, word editor or image editor history, stored data can be moved forward or backward by step(s).
--   New in-line operation functions `inc`, `mult`, `dev`, which are equivalent of `1 += 1`, `1 *= 2`, `i /= 2` in other programming languages.
+-   New in-line operation functions `inc`, `mult`, `divi`, which are equivalent of `1 += 1`, `1 *= 2`, `i /= 2` in other programming languages.
 
 ## Minor Change